Refactored surveyState and categoriesState
